Place secondary actions tests in appropriate folder
Also, fix lint errors
