+++ src/modules/list-secondary-actions/fixtures/list-secondary-actions.component.fixture.html
@@ -1,0 +1,9 @@
+<sky-list-toolbar>
+  <sky-list-secondary-actions>
+    <sky-list-secondary-action *ngIf="showOption">
+      <button type="button">
+        Option
+      </button>
+    </sky-list-secondary-action>
+  </sky-list-secondary-actions>
+</sky-list-toolbar>
