+++ src/modules/list-secondary-actions/fixtures/list-secondary-actions.component.fixture.ts
@@ -1,0 +1,20 @@
+import {
+  Component,
+  ViewChild
+} from '@angular/core';
+
+import {
+  SkyListSecondaryActionsComponent
+} from '../';
+
+@Component({
+  selector: 'sky-test-cmp',
+  template: require('./list-secondary-actions.component.fixture.html')
+})
+export class ListSecondaryActionsTestComponent {
+
+  public showOption: boolean = true;
+
+  @ViewChild(SkyListSecondaryActionsComponent)
+  public secondaryActions: SkyListSecondaryActionsComponent;
+}
