+++ src/modules/list-secondary-actions/fixtures/list-secondary-actions.component.visual-fixture.html
@@ -1,4 +1,4 @@
-<div id="screenshot-list-toolbar-secondary">
+<div id="screenshot-list-secondary-actions">
   <sky-list [data]="data">
     <sky-list-toolbar>
       <sky-list-toolbar-secondary-actions>
