--- src/modules/list-toolbar/fixtures/list-toolbar-secondary-actions.component.fixture.html
@@ -1,9 +1,0 @@
-<sky-list-toolbar>
-  <sky-list-toolbar-secondary-actions>
-    <sky-list-toolbar-secondary-action *ngIf="showOption">
-      <button type="button">
-        Option
-      </button>
-    </sky-list-toolbar-secondary-action>
-  </sky-list-toolbar-secondary-actions>
-</sky-list-toolbar>
