--- src/modules/list-toolbar/fixtures/list-toolbar-secondary-actions.component.fixture.ts
@@ -1,20 +1,0 @@
-import {
-  Component,
-  ViewChild
-} from '@angular/core';
-
-import {
-  SkyListToolbarSecondaryActionsComponent
-} from '../';
-
-@Component({
-  selector: 'sky-test-cmp',
-  template: require('./list-toolbar-secondary-actions.component.fixture.html')
-})
-export class ListToolbarSecondaryActionsTestComponent {
-
-  public showOption: boolean = true;
-
-  @ViewChild(SkyListToolbarSecondaryActionsComponent)
-  public secondaryActions: SkyListToolbarSecondaryActionsComponent;
-}
